refactor: removed all logic from index, leaving only exports
